// src/modules/company/create-company.dto.ts
export class CreateCompanyDto {


    code?: number;
  

    name?: string;
  

    address?: string;
  

    phone?: string;
  

    email?: string;
  

    website?: string;
  

    industry?: string;
  

    numberOfEmployees?: number;
  

    revenue?: number;
  

    description?: string;
  

    taxOffice?: string;  // Vergi Dairesi
  

    taxNumber?: string;  // Vergi Numarası
    
}